stock chart code refactor
